Crate xmpp_proto

XMPP Proto is the common and core structs and algorithm to interact between entities.

It offer everything to connect, manager, share stanza over a TcpStream.

Structs

Connection
Hold a Tcp/TlsStream and manage frames sent and receive over it.

XMPPCodec
A codec that will transform I/O into Event

XMPPTransport
Deal with the communication process based on a Tcp/Tls stream

XmppParser
XmppParser deals with incoming bytes. You can feed the parser with bytes and try to detect new event.

Enums

ConnectionState
Connection state information.

XMPPStream
Can be both Tcp or Tls. Own the TcpStream of a connection.